+++ src/components/Radio/index.ts
@@ -1,0 +1,9 @@
+import Radio from './Radio.vue'
+
+import { App } from 'vue'
+
+Radio.install = (app: App) => {
+  app.component(Radio.name, Radio)
+}
+
+export default Radio