+++ src/components/Tag/index.ts
@@ -1,0 +1,9 @@
+import Tag from './Tag.vue'
+
+import { App } from 'vue'
+
+Tag.install = (app: App) => {
+  app.component(Tag.name, Tag)
+}
+
+export default Tag