+++ src/demo/sections/ToastComponentExample.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <img
+      src="https://mk0culverduckguu41sc.kinstacdn.com/wp-content/uploads/2020/11/duck-animate-1-500x500.png"
+    />
+  </div>
+</template>
+
+<script lang="ts">
+import { Options, Vue } from 'vue-class-component'
+
+class Props {}
+
+@Options({})
+export default class ToastComponentExample extends Vue.with(Props) {}
+</script>
+
+<style lang="scss"></style>